LayoutTests/imported/w3c/web-platform-tests/css/css-color-adjust/meta/meta-color-scheme-single-value-in-body-expected.txt


FAIL Meta color-scheme in body does not apply. assert_equals: expected "normal" but got "auto"
